feat: 명령어 자동 등록 기능 추가
ClassGraph 라이브러리를 이용해서 Command 인터페이스를 이용하는 모든 클래스를 가져와 등록합니다.
