+++ src/main/kotlin/io/github/teamuselessplugin/punishment/interfaces/Command.kt
@@ -1,0 +1,5 @@
+package io.github.teamuselessplugin.punishment.interfaces
+
+interface Command {
+    fun push()
+}
